Attempts to improve performance along the Publish side by:

1. Removing all allocations during Int32 to UTF8 representation conversion.
2. Removing Buffer.BlockCopy on the fast-path for small fixed-length buffer (CRLF_BYTES).
3. Removing Array.Copy call for 0-length buffers.
4. Removing exceptions causing publish protocol buffer expansion.

```
|                        | Baseline          | Stupid Human Tricks | Delta           | Speedup       |
|-------------|----------|----------|--------|---------|-----------|---------|-------|--------|------|
|             | Count    | msgs/s   | kb/s   | msgs/s  | kb/s      | msgs/s  | kb/s  | msgs/s | kb/s |
| PubOnlyNo   | 10000000 | 2302097  | 0      | 4624634 | 0         | 2322537 | 0     | 2.01   |      |
| PubOnly8b   | 10000000 | 1982378  | 15487  | 3563483 | 27839     | 1581105 | 12352 | 1.80   | 1.80 |
| PubOnly32b  | 10000000 | 1664874  | 52027  | 2618534 | 81829     | 953660  | 29802 | 1.57   | 1.57 |
| PubOnly256b | 10000000 | 730606   | 182651 | 818758  | 204689    | 88152   | 22038 | 1.12   | 1.12 |
| PubOnly512b | 10000000 | 435323   | 217661 | 496377  | 248188    | 61054   | 30527 | 1.14   | 1.14 |
| PubOnly1k   | 1000000  | 233951   | 233951 | 276089  | 276089    | 42138   | 42138 | 1.18   | 1.18 |
| PubOnly4k   | 500000   | 57334    | 229336 | 67217   | 268868    | 9883    | 39532 | 1.17   | 1.17 |
| PubOnly8k   | 100000   | 28266    | 226128 | 32865   | 262920    | 4599    | 36792 | 1.16   | 1.16 |
| PubSubNo    | 10000000 | 603724   | 0      | 846889  | 0         | 243165  | 0     | 1.40   |      |
| PubSub8b    | 10000000 | 627404   | 4901   | 808974  | 6320      | 181570  | 1419  | 1.29   | 1.29 |
| PubSub32b   | 10000000 | 680405   | 21262  | 697522  | 21797     | 17117   | 535   | 1.03   | 1.03 |
| PubSub256b  | 10000000 | 386603   | 96650  | 462194  | 115548    | 75591   | 18898 | 1.20   | 1.20 |
| PubSub512b  | 500000   | 235251   | 117625 | 276339  | 138169    | 41088   | 20544 | 1.17   | 1.17 |
| PubSub1k    | 500000   | 112693   | 112693 | 155858  | 155858    | 43165   | 43165 | 1.38   | 1.38 |
| PubSub4k    | 500000   | 28192    | 112768 | 34501   | 138004    | 6309    | 25236 | 1.22   | 1.22 |
| PubSub8k    | 100000   | 15635    | 125080 | 17085   | 136680    | 1450    | 11600 | 1.09   | 1.09 |
| ReqReplNo   | 20000    | 4447     | 0      | 6233    | 0         | 1786    | 0     | 1.40   |      |
| ReqRepl8b   | 10000    | 4080     | 31     | 6452    | 50        | 2372    | 19    | 1.58   | 1.61 |
| ReqRepl32b  | 10000    | 3905     | 122    | 5931    | 185       | 2026    | 63    | 1.52   | 1.52 |
| ReqRepl256b | 5000     | 3912     | 978    | 5182    | 1295      | 1270    | 317   | 1.32   | 1.32 |
| ReqRepl512b | 5000     | 3396     | 1698   | 4840    | 2420      | 1444    | 722   | 1.43   | 1.43 |
| ReqRepl1k   | 5000     | 3767     | 3767   | 6070    | 6070      | 2303    | 2303  | 1.61   | 1.61 |
| ReqRepl4k   | 5000     | 2483     | 9932   | 4353    | 17412     | 1870    | 7480  | 1.75   | 1.75 |
| ReqRepl8k   | 5000     | 2106     | 16848  | 3330    | 26640     | 1224    | 9792  | 1.58   | 1.58 |
```
Average of ~1.38x speedup over master (pre #164).

The Int32 conversion routine fairs as follows against the existing code:
```
BenchmarkDotNet=v0.10.3.0, OS=Microsoft Windows NT 6.1.7601 Service Pack 1
Processor=Intel(R) Xeon(R) CPU E5-2620 v3 2.40GHzIntel(R) Xeon(R) CPU E5-2620 v3 2.40GHz, ProcessorCount=24
Frequency=2338378 Hz, Resolution=427.6469 ns, Timer=TSC
  [Host]     : Clr 4.0.30319.42000, 64bit RyuJIT-v4.7.2048.0
  DefaultJob : Clr 4.0.30319.42000, 64bit RyuJIT-v4.7.2048.0

| Method                     | Value   | Mean        | StdErr    | StdDev    | Scaled | Scaled-StdDev | Gen 0  | Allocated |
|----------------------------|---------|-------------|-----------|-----------|--------|---------------|--------|-----------|
| WriteStringToBuffer        | 8       | 101.4690 ns | 1.0240 ns | 7.5251 ns | 1      | 0             | 0.0039 | 32 B      |
| WriteInt32ToBufferInhuman2 | 8       | 2.8350 ns   | 0.0504 ns | 0.4217 ns | 0.03   | 0             | -      | 0 B       |
| WriteStringToBuffer        | 262144  | 108.9448 ns | 1.0875 ns | 4.8636 ns | 1      | 0             | 0.0053 | 40 B      |
| WriteInt32ToBufferInhuman2 | 262144  | 19.1210 ns  | 0.2537 ns | 2.5373 ns | 0.18   | 0.02          | -      | 0 B       |
| WriteStringToBuffer        | 1048576 | 121.7441 ns | 1.1012 ns | 4.2651 ns | 1      | 0             | 0.0031 | 40 B      |
| WriteInt32ToBufferInhuman2 | 1048576 | 20.7845 ns  | 0.2358 ns | 2.3465 ns | 0.17   | 0.02          | -      | 0 B       |
```